requirements.txt
absl-py==0.9.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
future==0.18.2
gast==0.3.3
gif==2.0.0
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-rl==0.4.2
kiwisolver==1.2.0
llvmlite==0.34.0
Markdown==3.2.2
matplotlib==3.3.1
numba==0.51.1
numpngw==0.0.8
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
torch==1.6.0+cu101
torchvision==0.7.0+cu101
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
